refactor(web): modularize main.tsx into App.tsx and hooks
